correcciones

+++ ProductManager.js
@@ -1,4 +1,5 @@
 import fs from "fs";
+import { title } from "process";
 
 export default class ProductManager {
   constructor() {
@@ -8,25 +9,26 @@ export default class ProductManager {
         if (fs.existsSync(this.path)) {
             const data = await fs.promises.readFile(this.path, "utf-8");
             const result = JSON.parse(data);
-            console.log(result);
             return result;
         } else {
             return [];
         }
     };
 
-    deletProduct = async(id) => {
+    deletProduct = async (id) => {
         const products = await this.consultarProductos();
 
-        try{
-            const productIndex = this.products.findIndex((product) => product.id === id);
+        const productIndex = this.products.findIndex((product) => product.id === id);
+        const productExists = products.find(element => element.id === id);
+
+        if (productExists) {
             products.splice(productIndex, 1);
             console.log("producto eliminado");
             await fs.promises.writeFile(this.path, JSON.stringify(products, null, "\t"));
             return products;
 
-        }catch(err){
-            console.log(`error: ${err}`);
+        } else {
+            return console.log("No se pudo eliminar el producto")
         }
 
     };
@@ -35,7 +37,6 @@ export default class ProductManager {
         const products = await this.consultarProductos();
  
         try{
-
             const product = products.find(element => element.id === id);
             return product ? product : null;
         }catch(err){
@@ -45,18 +46,28 @@ export default class ProductManager {
     };
 
     crearProducto = async (producto) => {
+
         const productos = await this.consultarProductos();
-        if (productos.length === 0) {
-            producto.id = 1;
+        const productExists = productos.find(element => element.code === producto.code);
+        if (producto.title.length === 0 || producto.description.length === 0 || producto.price.length === 0 || producto.thumbnail.length === 0 || producto.code.length === 0 || producto.stock.length === 0) {
+            return console.log("Hay un campo vacío");
         } else {
-            producto.id = productos[productos.length - 1].id + 1;
+
+            if (productExists) {
+                return console.log("El producto ya existe acá esta:", productExists);;
+            } else {
+                if (productos.length === 0) {
+                    producto.id = 1;
+                } else {
+                    producto.id = productos[productos.length - 1].id + 1;
+                }
+                productos.push(producto);
+                await fs.promises.writeFile(
+                    this.path,
+                    JSON.stringify(productos, null, "\t")
+                );
+            }
         }
-        productos.push(producto);
-        await fs.promises.writeFile(
-            this.path,
-            JSON.stringify(productos, null, "\t")
-        );
-        return producto;
     };
